[WIP] Case study development
