+++ app/assets/styles/partials/_base.scss
@@ -111,23 +111,21 @@ body {
 h1 {
 	font-family: $titles-type;
 	font-size: 4.5rem;
-	font-weight: 300;
+	font-weight: 400;
 	line-height: 5.3rem;
 	margin: unset;
-	padding-bottom: $title-padding;
+	margin-bottom: $title-padding;
 	/*@include column(8/12);*/
 	display: inline-block;
-
-
 }
 
 
 h2{
 	font-family: $titles-type;
 	font-size: 3rem;
-	font-weight: 300;
+	font-weight: 400;
 	line-height: 1.3;
-	padding-bottom: $title-padding;
+	margin-bottom: $title-padding;
 	display: inline-block;
 }
 
@@ -135,7 +133,7 @@ h2{
 h3{
 	font-family: $titles-type;
 	font-size: 1.5rem;
-	font-weight: 300;
+	font-weight: 400;
 	line-height: 1.8rem;
 	display: inline-block;
 }
